break up uri module into sub modules
